pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "windows-mcp-server"
version = "0.1.0"
description = "MCP server for Windows PC automation and control"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Windows MCP", email = "windows-mcp@example.com" }
]
dependencies = [
    "mcp>=0.9.0",
    "pillow>=10.0.0",
    "pyautogui>=0.9.54",
    "pywin32>=306",
    "psutil>=5.9.0",
    "mss>=9.0.1",
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.0",
    "uiautomation>=2.0.18",
    "tabulate>=0.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.5.0"
]

[tool.setuptools]
packages = ["windows_mcp", "windows_mcp.desktop", "windows_mcp.tree"]

[project.scripts]
windows-mcp = "windows_mcp.server:main"
